test: add initial test for orchestrator module with more to follow
